GRIDNET Server now accepts custom submit files
Job files can now be entered manually. The user still needs to select
the grid (if any in particular) the job needs to be run on and the files
it requires, but all other options are left to the user.
The task will be shown as "custom". There can be no guarantee that the
task will run on an adequate node.