Added extra info to form submission
When uploading large files it may have been unclear whether or not a 
job was actually being accepted. A notification has been added to inform
the user that the upload is pending. Additionaly, extra AJAX error
handling has been added to inform the user if a server error occured
during submission.